+++ src/neon/yuv_to_rgba.rs
@@ -50,7 +50,7 @@ pub unsafe fn neon_yuv_to_rgba_row<const DESTINATION_CHANNELS: u8, const SAMPLIN
     let v_g_coeff_2 = vdupq_n_s16(-1i16 * transform.g_coeff_2 as i16);
     let v_alpha = vdupq_n_u8(255u8);
 
-    while cx + 16 < width as usize {
+    while cx + 16 < width {
         let y_values = vsubq_u8(vld1q_u8(y_ptr.add(y_offset + cx)), y_corr);
 
         let u_high_u8: uint8x8_t;
